# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Avalanche P-Chain and EVM Indexer API\n\n## Overview\nThis API provides access to indexed Avalanche blockchain data including:\n- **Data API** (`/api/v1/data/*`): Blocks, transactions, subnets, validators, P-chain data\n- **Metrics API** (`/api/v1/metrics/*`): Fee statistics, chain metrics, time series data\n\n## Rate Limiting\nAll endpoints are rate limited to 60 requests/minute per IP with a burst of 10.",
        "title": "Icicle API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/data/pchain/tx-types": {
            "get": {
                "description": "Get a list of P-Chain transaction types with counts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data - P-Chain"
                ],
                "summary": "List P-Chain transaction types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/data/pchain/txs": {
            "get": {
                "description": "Get a paginated list of P-Chain transactions with optional filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data - P-Chain"
                ],
                "summary": "List P-Chain transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by transaction type",
                        "name": "tx_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by subnet ID",
                        "name": "subnet_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Number of results (max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api.PChainTx"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/api.Meta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/data/pchain/txs/{txId}": {
            "get": {
                "description": "Get details for a specific P-Chain transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data - P-Chain"
                ],
                "summary": "Get P-Chain transaction by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "txId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.PChainTx"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/api.ErrorCode"
                },
                "details": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "retry_after": {
                    "type": "integer"
                }
            }
        },
...
        "api.ErrorCode": {
            "type": "string",
            "enum": [
                "INVALID_PARAMETER",
                "NOT_FOUND",
                "INTERNAL_ERROR",
                "RATE_LIMITED",
                "VALIDATION_FAILED",
                "DATABASE_ERROR"
            ],
            "x-enum-varnames": [
                "ErrInvalidParameter",
                "ErrNotFound",
                "ErrInternalError",
                "ErrRateLimited",
                "ErrValidationFailed",
                "ErrDatabaseError"
            ]
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.APIError"
                }
            }
        },
...
        "api.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "api.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.PChainTx": {
            "type": "object",
            "properties": {
                "block_number": {
                    "type": "integer",
                    "example": 12345678
                },
                "block_time": {
                    "type": "string"
                },
                "tx_data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "tx_id": {
                    "type": "string",
                    "example": "2ZW6HUePB..."
                },
                "tx_type": {
                    "type": "string",
                    "example": "ConvertSubnetToL1Tx"
                }
            }
        },
        "api.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/api.Meta"
                }
            }
        },
...
    },
...
}